# Optimise `SymmetricDifference`