## Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Bunch of optimisations of `set.SymmetricDifference` from the somewhat naive first implementation, results in a very significant performance improvement:

```plaintext
goos: darwin
goarch: amd64
pkg: github.com/FollowTheProcess/collections/set
cpu: Intel(R) Core(TM) i5-8257U CPU @ 1.40GHz
                      │  before.txt  │              after.txt              │
                      │    sec/op    │   sec/op     vs base                │
SymmetricDifference-8   208.18µ ± 1%   85.84µ ± 0%  -58.77% (p=0.000 n=30)

                      │  before.txt  │              after.txt               │
                      │     B/op     │     B/op      vs base                │
SymmetricDifference-8   93.67Ki ± 0%   21.44Ki ± 0%  -77.11% (p=0.000 n=30)

                      │  before.txt  │             after.txt              │
                      │  allocs/op   │ allocs/op   vs base                │
SymmetricDifference-8   150.000 ± 0%   8.000 ± 0%  -94.67% (p=0.000 n=30)
```